+++ demo/node/src/staging.rs
@@ -3,9 +3,8 @@ use crate::chain_spec::*;
 use authority_selection_inherents::CommitteeMember;
 use partner_chains_demo_runtime::{
 	AccountId, AuraConfig, BalancesConfig, BeefyConfig, BridgeConfig, GovernedMapConfig,
-	GrandpaConfig, NativeTokenManagementConfig, RuntimeGenesisConfig,
-	SessionCommitteeManagementConfig, SessionConfig, SidechainConfig, SudoConfig, SystemConfig,
-	TestHelperPalletConfig,
+	GrandpaConfig, RuntimeGenesisConfig, SessionCommitteeManagementConfig, SessionConfig,
+	SidechainConfig, SudoConfig, SystemConfig, TestHelperPalletConfig,
 };
 use sc_service::ChainType;
 use sidechain_domain::*;
@@ -133,15 +132,19 @@ pub fn staging_genesis(
 		},
 		transaction_payment: Default::default(),
 		session: SessionConfig {
-			initial_validators: initial_authorities
+			keys: initial_authorities
 				.iter()
 				.map(|authority_keys| {
-					(authority_keys.cross_chain.clone().into(), authority_keys.session.clone())
+					(
+						authority_keys.cross_chain.clone().into(),
+						authority_keys.cross_chain.clone().into(),
+						authority_keys.session.clone(),
+					)
 				})
 				.collect(),
+			non_authority_keys: Default::default(),
 		},
 		sidechain: SidechainConfig { genesis_utxo, ..Default::default() },
-		pallet_session: Default::default(),
 		session_committee_management: SessionCommitteeManagementConfig {
 			initial_authorities: initial_authorities
 				.into_iter()