+++ demo/node/src/template_chain_spec.rs
@@ -1,8 +1,8 @@
 use crate::chain_spec::*;
 use partner_chains_demo_runtime::{
 	AuraConfig, BalancesConfig, BeefyConfig, BridgeConfig, GovernedMapConfig, GrandpaConfig,
-	NativeTokenManagementConfig, RuntimeGenesisConfig, SessionCommitteeManagementConfig,
-	SessionConfig, SidechainConfig, SudoConfig, SystemConfig, TestHelperPalletConfig,
+	RuntimeGenesisConfig, SessionCommitteeManagementConfig, SessionConfig, SidechainConfig,
+	SudoConfig, SystemConfig, TestHelperPalletConfig,
 };
 use sc_service::ChainType;
 
@@ -28,11 +28,11 @@ pub fn chain_spec() -> Result<ChainSpec, envy::Error> {
 		},
 		transaction_payment: Default::default(),
 		session: SessionConfig {
-			// Initial validators are meant to be updated in the chain spec file, so it is empty here.
-			initial_validators: vec![],
+			// Keys are meant to be updated in the chain spec file, so it is empty here.
+			keys: vec![],
+			non_authority_keys: Default::default(),
 		},
 		sidechain: SidechainConfig { genesis_utxo, ..Default::default() },
-		pallet_session: Default::default(),
 		session_committee_management: SessionCommitteeManagementConfig {
 			// Same as SessionConfig
 			initial_authorities: vec![],