+++ dev/local-environment/setup.sh
@@ -5,6 +5,7 @@ PARTNER_CHAINS_NODE_IMAGE="partner-chains-node:beefy"
 # PARTNER_CHAINS_NODE_IMAGE="partner-chains-node:latest"
 CARDANO_IMAGE="ghcr.io/intersectmbo/cardano-node:10.5.1"
 DBSYNC_IMAGE="ghcr.io/intersectmbo/cardano-db-sync:13.6.0.4"
+DOLOS_IMAGE="ghcr.io/txpipe/dolos:1.0.0-beta.5"
 OGMIOS_IMAGE="cardanosolutions/ogmios:v6.13.0"
 POSTGRES_IMAGE="postgres:17.2"
 TESTS_IMAGE="python:3.12-slim"
@@ -124,6 +125,23 @@ function validate_memory_limit() {
   done
 }
 
+configure_dolos() {
+  echo "===== DOLOS CONFIGURATION ========"
+  read -p "Do you want to set a non-default gRPC port for Dolos? (Will default to 50051) (Y/N): " set_dolos_grpc_port
+  if [[ $set_dolos_grpc_port == [Yy]* ]]; then
+    dolos_grpc_port=$(validate_port "Enter gRPC port for Dolos: ")
+  else
+    dolos_grpc_port=50051
+  fi
+  read -p "Do you want to set a non-default miniBF port for Dolos? (Will default to 3000) (Y/N): " set_dolos_minibf_port
+  if [[ $set_dolos_minibf_port == [Yy]* ]]; then
+    dolos_minibf_port=$(validate_port "Enter gRPC port for Dolos: ")
+  else
+    dolos_minibf_port=3000
+  fi
+  echo
+}
+
 configure_ogmios() {
   echo "===== OGMIOS CONFIGURATION ========"
   read -p "Do you want to set a non-default port for Ogmios? (Will default to 1337) (Y/N): " set_ogmios_port
@@ -241,6 +259,8 @@ configure_env() {
         cat <<EOF >.env
 POSTGRES_PORT=5432
 POSTGRES_PASSWORD=$db_password
+DOLOS_MINIBF_PORT=3000
+DOLOS_GRPC_PORT=50051
 OGMIOS_PORT=1337
 CPU_PARTNER_CHAINS_NODE=0.000
 MEM_PARTNER_CHAINS_NODE=1000G
@@ -250,6 +270,8 @@ CPU_POSTGRES=0.000
 MEM_POSTGRES=1000G
 CPU_DBSYNC=0.000
 MEM_DBSYNC=1000G
+CPU_DOLOS=0.000
+MEM_DOLOS=1000G
 CPU_OGMIOS=0.000
 MEM_OGMIOS=1000G
 EOF
@@ -257,6 +279,8 @@ EOF
         cat <<EOF >.env
 POSTGRES_PORT=$db_port
 POSTGRES_PASSWORD=$db_password
+DOLOS_MINIBF_PORT=$dolos_minibf_port
+DOLOS_GRPC_PORT=$dolos_grpc_port
 OGMIOS_PORT=$ogmios_port
 CPU_PARTNER_CHAINS_NODE=$CPU_PARTNER_CHAINS_NODE
 MEM_PARTNER_CHAINS_NODE=$MEM_PARTNER_CHAINS_NODE
@@ -266,6 +290,8 @@ CPU_POSTGRES=$cpu_postgres
 MEM_POSTGRES=$mem_postgres
 CPU_DBSYNC=$cpu_dbsync
 MEM_DBSYNC=$mem_dbsync
+CPU_DOLOS=$cpu_dolos
+MEM_DOLOS=$mem_dolos
 CPU_OGMIOS=$cpu_ogmios
 MEM_OGMIOS=$mem_ogmios
 EOF
@@ -274,6 +300,7 @@ EOF
     cat <<EOF >>.env
 CARDANO_IMAGE=$CARDANO_IMAGE
 DBSYNC_IMAGE=$DBSYNC_IMAGE
+DOLOS_IMAGE=$DOLOS_IMAGE
 OGMIOS_IMAGE=$OGMIOS_IMAGE
 POSTGRES_IMAGE=$POSTGRES_IMAGE
 TESTS_IMAGE=$TESTS_IMAGE
@@ -295,7 +322,8 @@ choose_deployment_option() {
     echo "3) Include Cardano testnet, Ogmios, DB-Sync and Postgres"
     echo "4) Deploy a single Partner Chains node with network_mode: "host" for external connections (adjust partner-chains-external-node.txt before running this script)"
     echo "5) Deploy a 3 node Partner Chain network using wizard"
-    read -p "Enter your choice (1/2/3/4/5): " deployment_option
+    echo "6) Include Cardano testnet, Ogmios, and Dolos"
+    read -p "Enter your choice (1/2/3/4/5/6): " deployment_option
   else
     deployment_option=0
   fi
@@ -323,11 +351,12 @@ create_docker_compose() {
         cat ./modules/ogmios.txt >> docker-compose.yml
         ;;
       3)
-        echo -e "Including Cardano testnet, Ogmios, DB-Sync, and Postgres services.\n"
+        echo -e "Including Cardano testnet, Ogmios, Dolos, DB-Sync, and Postgres services.\n"
         cat ./modules/cardano.txt >> docker-compose.yml
         cat ./modules/ogmios.txt >> docker-compose.yml
         cat ./modules/db-sync.txt >> docker-compose.yml
         cat ./modules/postgres.txt >> docker-compose.yml
+        cat ./modules/dolos.txt >> docker-compose.yml
         ;;
       4)
         echo -e "Including all services with external partner chain node.\n"
@@ -335,6 +364,7 @@ create_docker_compose() {
         cat ./modules/ogmios.txt >> docker-compose.yml
         cat ./modules/db-sync.txt >> docker-compose.yml
         cat ./modules/postgres.txt >> docker-compose.yml
+        cat ./modules/dolos.txt >> docker-compose.yml
         cat ./modules/partner-chains-external-node.txt >> docker-compose.yml
         cat ./modules/partner-chains-setup.txt >> docker-compose.yml
         ;;
@@ -344,7 +374,14 @@ create_docker_compose() {
         cat ./modules/ogmios.txt >> docker-compose.yml
         cat ./modules/db-sync.txt >> docker-compose.yml
         cat ./modules/postgres.txt >> docker-compose.yml
+        cat ./modules/dolos.txt >> docker-compose.yml
         cat ./modules/partner-chains-wizard.txt >> docker-compose.yml
+        ;;
+      6)
+        echo -e "Including Cardano testnet, Ogmios, and Dolos services.\n"
+        cat ./modules/cardano.txt >> docker-compose.yml
+        cat ./modules/ogmios.txt >> docker-compose.yml
+        cat ./modules/dolos.txt >> docker-compose.yml
         ;;
       0)
         echo -e "Including all services.\n"
@@ -352,6 +389,7 @@ create_docker_compose() {
         cat ./modules/ogmios.txt >> docker-compose.yml
         cat ./modules/db-sync.txt >> docker-compose.yml
         cat ./modules/postgres.txt >> docker-compose.yml
+        cat ./modules/dolos.txt >> docker-compose.yml
         cat ./modules/partner-chains-nodes.txt >> docker-compose.yml
         cat ./modules/partner-chains-setup.txt >> docker-compose.yml
         ;;
@@ -381,11 +419,11 @@ parse_arguments() {
                 shift
                 ;;
             -d|--deployment-option)
-                if [[ -n "$2" && "$2" =~ ^[1-5]$ ]]; then
+                if [[ -n "$2" && "$2" =~ ^[1-6]$ ]]; then
                     deployment_option="$2"
                     shift 2
                 else
-                    echo "Error: Invalid deployment option '$2'. Valid options are 1, 2, 3, 4 or 5."
+                    echo "Error: Invalid deployment option '$2'. Valid options are 1, 2, 3, 4, 5 or 6."
                     exit 1
                 fi
                 ;;
@@ -456,6 +494,7 @@ main() {
         detect_os "interactive"
         backup_files "interactive"
         configure_postgres "interactive"
+        configure_dolos
         configure_ogmios
         resource_limits_setup