+++ e2e-tests/tests/conftest.py
@@ -264,17 +264,8 @@ def load_config(blockchain, nodes_env, ci_run, node_host, node_port, deployment_
     assert os.path.isfile(blockchain_config_path), f"Config file not found {blockchain_config_path}"
     blockchain_config = OmegaConf.load(blockchain_config_path)
 
-    stack_config_path = f"{os.getcwd()}/config/{blockchain}/{nodes_env}_stack.json"
-    assert os.path.isfile(stack_config_path), f"Config file not found {stack_config_path}"
-    stack_config = OmegaConf.load(stack_config_path)
-
     schema = OmegaConf.structured(ApiConfig)
-    config: ApiConfig = OmegaConf.merge(schema, default_config, blockchain_config, stack_config)
-
-    ci_config_path = f"{os.getcwd()}/config/{blockchain}/{nodes_env}-ci.json"
-    if ci_run and os.path.isfile(ci_config_path):
-        ci_config = OmegaConf.load(ci_config_path)
-        config = OmegaConf.merge(config, ci_config)
+    config: ApiConfig = OmegaConf.merge(schema, default_config, blockchain_config)
 
     # command line arguments that override config values
     if node_host: