+++ toolkit/data-sources/db-sync/src/mc_hash/mod.rs
@@ -1,8 +1,6 @@
 //! Db-Sync data source used by Partner Chain MC Reference feature
-use crate::{
-	block::BlockDataSourceImpl,
-	metrics::{McFollowerMetrics, observed_async_trait},
-};
+use crate::block::BlockDataSourceImpl;
+use partner_chains_data_source_metrics::{McFollowerMetrics, observed_async_trait};
 use sidechain_domain::{MainchainBlock, McBlockHash};
 use sidechain_mc_hash::McHashDataSource;
 use sp_timestamp::Timestamp;