+++ toolkit/data-sources/db-sync/src/sidechain_rpc/mod.rs
@@ -1,8 +1,7 @@
 //! Db-Sync data source used by Partner Chain Json RPC
-use crate::{
-	block::BlockDataSourceImpl, metrics::McFollowerMetrics, metrics::observed_async_trait,
-};
+use crate::block::BlockDataSourceImpl;
 use pallet_sidechain_rpc::SidechainRpcDataSource;
+use partner_chains_data_source_metrics::{McFollowerMetrics, observed_async_trait};
 use sidechain_domain::MainchainBlock;
 use std::sync::Arc;