+++ toolkit/partner-chains-cli/src/create_chain_spec/tests.rs
@@ -13,6 +13,7 @@ impl PartnerChainRuntime for MockRuntime {
 		config: &super::CreateChainSpecConfig<TestSessionKeys>,
 	) -> serde_json::Value {
 		serde_json::json!({
+			"bootnodes": config.bootnodes,
 			"session":config.pallet_partner_chains_session_config::<MockRuntime>(),
 			"sessionCommitteeManagement": config.pallet_session_validator_management_config::<MockRuntime>(),
 			"sidechain": config.pallet_sidechain_config::<MockRuntime>(SlotsPerEpoch(13)),
@@ -95,6 +96,7 @@ If you are a validator, you can obtain the chain configuration file from the gov
 
 fn test_config_content() -> serde_json::Value {
 	serde_json::json!({
+		"bootnodes": ["test_bootnode"],
 		"chain_parameters": chain_parameters_json(),
 		"initial_permissioned_candidates": [
 			{
@@ -131,6 +133,7 @@ fn test_config_content_without_initial_permissioned_candidates() -> serde_json::
 
 fn test_config_content_with_empty_initial_permissioned_candidates() -> serde_json::Value {
 	serde_json::json!({
+		"bootnodes": [],
 		"chain_parameters": chain_parameters_json(),
 		"initial_permissioned_candidates": [],
 		"cardano_addresses": cardano_addresses_json(),
@@ -215,9 +218,11 @@ fn show_initial_permissioned_candidates() -> MockIO {
 fn generated_chain_spec() -> serde_json::Value {
 	serde_json::json!(
 		{
+			"bootnodes": ["test_bootnode"],
 			"session": {
-				"initialValidators": [
+				"keys": [
 					[
+						"5C7C2Z5sWbytvHpuLTvzKunnnRwQxft1jiqrLD5rhucQ5S9X",
 						"5C7C2Z5sWbytvHpuLTvzKunnnRwQxft1jiqrLD5rhucQ5S9X",
 						{
 							"aura": "5GrwvaEF5zXb26Fz9rcQpDWS57CtERHpNehXCPcNoHGKutQY",
@@ -226,12 +231,14 @@ fn generated_chain_spec() -> serde_json::Value {
 					],
 					[
 						"5DVskgSC9ncWQpxFMeUn45NU43RUq93ByEge6ApbnLk6BR9N",
+						"5DVskgSC9ncWQpxFMeUn45NU43RUq93ByEge6ApbnLk6BR9N",
 						{
 							"aura": "5FHneW46xGXgs5mUiveU4sbTyGBzmstUspZC92UhjJM694ty",
 							"grandpa": "5GoNkf6WdbxCFnPdAnYYQyCjAKPJgLNxXwPjwTh6DGg6gN3E"
 						}
 					]
-				]
+				],
+				"nonAuthorityKeys": []
 			},
 			"sessionCommitteeManagement": {
 				"initialAuthorities": [