+++ changelog.md
@@ -7,8 +7,14 @@ This changelog is based on [Keep A Changelog](https://keepachangelog.com/en/1.1.
 ## Changed
 
 * Updated partner-chains-smart-contracts (raw-scripts) dependency to v8.2.0. Not breaking.
+* `SessionManager` and `ShouldEndSession` implementations of `pallet-session-validator-management` rotate one additional
+session in order to make `pallet_session` use authorities with less delay.
 
 ## Removed
+
+* `pallet-partner-chains-session` has been removed. Partner Chains should use only the stock Substrate session pallet
+* `PalletSessionSupport` type provided by `pallet-session-validator-management`. The `SessionManager` and `ShouldEndSession`
+  implementations were moved directly to the `Pallet` type instead.
 
 ## Fixed